Refactor rate limit error handling across components
- Introduced a centralized utility function `isRateLimitError` to streamline rate limit error detection in ProjectManager, ChatControls, and Docs4LLMParser.
- Removed redundant error checking logic from multiple locations, enhancing code maintainability.
- Added a new utility function `extractRetryTime` for better handling of retry time extraction from error messages.
