+++ src/components/chat-components/ChatControls.tsx
@@ -12,6 +12,7 @@ import { navigateToPlusPage, useIsPlusUser } from "@/plusUtils";
 import VectorStoreManager from "@/search/vectorStoreManager";
 import { updateSetting, useSettingsValue } from "@/settings/model";
 import { Docs4LLMParser } from "@/tools/FileParserManager";
+import { isRateLimitError } from "@/utils/rateLimitUtils";
 import { DropdownMenu, DropdownMenuTrigger } from "@radix-ui/react-dropdown-menu";
 import {
   AlertTriangle,
@@ -79,14 +80,7 @@ export async function reloadCurrentProject() {
     logError("Error reloading project context:", error);
 
     // Check if this is a rate limit error and let the FileParserManager notice handle it
-    const errorMessage = error?.message || error?.toString() || "";
-    const isRateLimit =
-      errorMessage.includes("Request rate limit exceeded") ||
-      errorMessage.includes("RATE_LIMIT_EXCEEDED") ||
-      errorMessage.includes("429") ||
-      error?.status === 429;
-
-    if (!isRateLimit) {
+    if (!isRateLimitError(error)) {
       new Notice("Failed to reload project context. Check console for details.");
     }
     // If it's a rate limit error, don't show generic failure message - let the rate limit notice show
@@ -135,14 +129,7 @@ export async function forceRebuildCurrentProjectContext() {
         logError("Error force rebuilding project context:", error);
 
         // Check if this is a rate limit error and let the FileParserManager notice handle it
-        const errorMessage = error?.message || error?.toString() || "";
-        const isRateLimit =
-          errorMessage.includes("Request rate limit exceeded") ||
-          errorMessage.includes("RATE_LIMIT_EXCEEDED") ||
-          errorMessage.includes("429") ||
-          error?.status === 429;
-
-        if (!isRateLimit) {
+        if (!isRateLimitError(error)) {
           new Notice("Failed to force rebuild project context. Check console for details.");
         }
         // If it's a rate limit error, don't show generic failure message - let the rate limit notice show
